chore(pytest) - Enable subprocess tests to use pytest
This change is being done to allow subprocess tests to use fixtures.
Prior to this change, unittest classes with methods marked with the
`TracerTestCase.run_in_subprocess()` method were unable to use pytest
fixtures for two reasons:

1. Subprocesses were invoked via unittest, which requires special
   handling for including fixtures
2. The special handling to include fixtures from (1) failed to apply the
   fixtures to the methods, because of the method signature ambiguation
   applied by the run_in_test wrapper.

This change fixes the issue by removing the need for the special
handling in (1) by allowing subprocess tests to, optionally, be invoked
via pytest.